third_party/WebKit/LayoutTests/svg/animations/animateMotion-still-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests


Test animations that only express an offset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().x is 200
PASS rootSVGElement.getBBox().x is 200
PASS rootSVGElement.getBBox().x is 0
PASS rootSVGElement.getBBox().x is 0
PASS successfullyParsed is true

TEST COMPLETE
